[MM-36984] - Update Cypress test for onboarding flow
