+++ e2e/cypress/integration/onboarding/login_page_link_account_creation_spec.js
@@ -8,15 +8,15 @@
 // ***************************************************************
 
 // Stage: @prod
-// Group: @enterprise @onboarding @cloud_only
+// Group: @enterprise @onboarding
 
-import * as TIMEOUTS from '../../../../fixtures/timeouts';
-import {generateRandomUser} from '../../../../support/api/user';
+import * as TIMEOUTS from '../../fixtures/timeouts';
+import {generateRandomUser} from '../../support/api/user';
 import {
     getWelcomeEmailTemplate,
     reUrl,
     verifyEmailBody,
-} from '../../../../utils';
+} from '../../utils';
 
 describe('Onboarding', () => {
     let siteName;
