+++ e2e/cypress/integration/onboarding/set_profile_and_team_spec.js
